refactor(core): improve testability and extend functionality across services
- Refactored RikImportService and RikProcessService to decouple logic and improve testability.
- Added PollingPlaceService for better separation of responsibilities.
- Introduced updateVotesNumbers method across all voter-related entities to simplify future vote recalculations.
- Created custom exception package with three domain-specific exception classes.
- Added validation constraints to relevant domain classes (e.g., @NotNull, @Size).

+++ src/main/java/com/discord/Projekat_Izbori/repositories/FinalResultsRepository.java
@@ -3,5 +3,5 @@ package com.discord.Projekat_Izbori.repositories;
 import com.discord.Projekat_Izbori.models.FinalResults;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FinalResultsRepository extends JpaRepository<FinalResults, Long> {
+public interface FinalResultsRepository extends JpaRepository<FinalResults, Integer> {
 }
