+++ src/main/java/com/discord/Projekat_Izbori/repositories/IrregularityRepositiry.java
@@ -3,5 +3,5 @@ package com.discord.Projekat_Izbori.repositories;
 import com.discord.Projekat_Izbori.models.Irregularity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IrregularityRepositiry extends JpaRepository<Irregularity, Long> {
+public interface IrregularityRepositiry extends JpaRepository<Irregularity, Integer> {
 }
