+++ src/main/java/com/discord/Projekat_Izbori/repositories/PoliticalPartyRepository.java
@@ -3,5 +3,5 @@ package com.discord.Projekat_Izbori.repositories;
 import com.discord.Projekat_Izbori.models.PoliticalParty;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PoliticalPartyRepository extends JpaRepository<PoliticalParty, Long> {
+public interface PoliticalPartyRepository extends JpaRepository<PoliticalParty, Integer> {
 }
