+++ src/main/java/com/discord/Projekat_Izbori/repositories/PoliticalPartyResultRepository.java
@@ -3,5 +3,5 @@ package com.discord.Projekat_Izbori.repositories;
 import com.discord.Projekat_Izbori.models.PoliticalPartyResult;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PoliticalPartyResultRepository extends JpaRepository<PoliticalPartyResult, Long> {
+public interface PoliticalPartyResultRepository extends JpaRepository<PoliticalPartyResult, Integer> {
 }
